+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Remove redundant alembic upgrade that can cause upgrade errors
  - Fix moving Conditional code during upgrade
+ - Fix MQTT Input ([#685](https://github.com/kizniche/mycodo/issues/685))
 
 # TODO Add db revision that saves conditional code to user code directory